Moved and add tests to properly directory
