+++ .github/workflows/build.yml
@@ -40,6 +40,7 @@ jobs:
                 uses: shivammathur/setup-php@v2
                 with:
                     php-version: "${{ matrix.php }}"
+                    extensions: mongodb-1.21.0
                     coverage: none
                     tools: 'composer:v2, flex'
 
@@ -183,6 +184,7 @@ jobs:
                 uses: shivammathur/setup-php@v2
                 with:
                     php-version: "${{ matrix.php }}"
+                    extensions: mongodb-1.21.0
                     coverage: none
 
             -
